requirements.txt
amqp==5.0.9
app==0.0.1
asgiref==3.5.0
Babel==2.9.1
backports.zoneinfo==0.2.1
bcrypt==3.2.2
billiard==3.6.4.0
boto3==1.24.7
botocore==1.27.7
celery==5.2.3
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.10
class-registry==2.1.2
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
Deprecated==1.2.13
Django==3.2.11
django-braces==1.15.0
django-celery-beat==2.2.1
django-cors-headers==3.11.0
django-filter==21.1
django-money==2.1.1
django-phonenumber-field==6.1.0
django-storages==1.12.3
django-taggit==2.1.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
docker==5.0.3
drf-yasg==1.20.0
example==0.1.0
filters==1.3.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==1.0.0
jsonfield==3.1.0
kombu==5.2.3
MarkupSafe==2.0.1
packaging==21.3
paramiko==2.11.0
phonenumbers==8.12.48
phonenumberslite==8.12.48
pi==0.1.2
Pillow==9.1.0
prompt-toolkit==3.0.26
psycopg2-binary==2.9.3
py-moneyed==1.2
pycparser==2.21
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
redis==4.1.2
regex==2022.3.2
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
stripe==2.76.0
tblib==1.7.0
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.2.3
wrapt==1.13.3
